fix(security): resolve CodeQL alerts — workflow permissions, same-origin redirect guard, size_t cast
- privacy-scan.yml: add explicit least-privilege permissions block (actions/missing-workflow-permissions)
- Login.vue: sanitizeRedirect now resolves against window.location.origin and rejects cross-origin results (js/client-side-unvalidated-url-redirection x2)
- cuda.cpp: widen row_pitch*height to size_t before resize (cpp/integer-multiplication-cast-to-long)

+++ src/platform/linux/cuda.cpp
@@ -170,7 +170,7 @@ namespace cuda {
       img.row_pitch = img.width * img.pixel_pitch;
 
       std::vector<std::uint8_t> image_data;
-      image_data.resize(img.row_pitch * img.height);
+      image_data.resize(static_cast<std::size_t>(img.row_pitch) * img.height);
 
       img.data = image_data.data();
